% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{as.data.frame.survfit}
\alias{as.data.frame.survfit}
\title{Convert \link[survival]{survfit.object} to \link[base]{data.frame}}
\usage{
\method{as.data.frame}{survfit}(x, ..., times, rm01 = FALSE)
}
\arguments{
\item{x}{\link[survival]{survfit.object}.
\code{'survfitcox'} object returned from
\code{survival:::survfit.coxph} is also supported, as of 2022-09-19}

\item{...}{additional parameters of \link[base]{as.data.frame.list}}

\item{times}{(optional) \link[base]{numeric} vector}

\item{rm01}{\link[base]{logical} scalar, whether to remove survival rates (and confidence bands)
being 0 or 1, as some y-axis transformation may yield \code{NA} or \code{Inf} values.  Default \code{FALSE}.}
}
\value{
Function \code{\link[=as.data.frame.survfit]{as.data.frame.survfit()}} returns a \link[base]{data.frame}
}
\description{
..
}
\details{
A column named \code{'strata'}
of all-equal values \code{'all_subjects'}
for \link[survival]{survfit.object} without a strata.
}
\seealso{
\code{survival:::summary.survfit}
}
