% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pROC_ext_vanilla.R
\name{autoplot.roc}
\alias{autoplot.roc}
\title{autoplot.roc}
\usage{
\method{autoplot}{roc}(object, ...)
}
\arguments{
\item{object}{roc}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\note{
Function \code{pROC:::plot.roc.roc} does \emph{not} return a \link[ggplot2]{ggplot} figure.
}
