% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_roc.R
\name{autoplot.survival_roc}
\alias{autoplot.survival_roc}
\title{autoplot.survival_roc}
\usage{
\method{autoplot}{survival_roc}(object, ...)
}
\arguments{
\item{object}{\code{'survival_roc'} object}

\item{...}{..}
}
\description{
..
}
