% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pROC_ext_vanilla.R
\name{md_.roc}
\alias{md_.roc}
\title{md_.roc}
\usage{
\method{md_}{roc}(x, xnm, ...)
}
\arguments{
\item{x}{\link[pROC]{roc} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\keyword{internal}
