% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{roc.tzh-package}
\alias{roc.tzh}
\alias{roc.tzh-package}
\title{roc.tzh}
\description{
roc.tzh
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
