% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_roc.R
\name{Sprintf.survival_roc}
\alias{Sprintf.survival_roc}
\title{Sprintf.survival_roc}
\usage{
Sprintf.survival_roc(model, ...)
}
\arguments{
\item{model}{\code{'survival_roc'} object}

\item{...}{..}
}
\description{
..
}
