% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_roc.R
\name{survival_roc}
\alias{survival_roc}
\alias{riskset_roc}
\title{Survival ROC and Risk Set ROC}
\usage{
survival_roc(formula, data, predict.time, span = 0.25 * NROW(data)^(-0.2), ...)

riskset_roc(formula, data, predict.time, plot = FALSE, ...)
}
\arguments{
\item{formula}{..}

\item{data}{..}

\item{predict.time, span, ...}{additional parameters of
\link[survivalROC]{survivalROC} and \link[risksetROC]{risksetROC}}

\item{plot}{must be \code{FALSE}, to suppress the plot from \link[risksetROC]{risksetROC}}
}
\description{
Convert the returned objects of \link[survivalROC]{survivalROC} and \link[risksetROC]{risksetROC}
into \code{'roc'} objects (as returned by \link[pROC]{roc}),
to take advantages of S3 methods such as \link[pROC]{plot.roc} and \code{\link[=autoplot.roc]{autoplot.roc()}}.
}
\examples{
?survivalROC::survivalROC
data(mayo, package = 'survivalROC')
library(survival)
m1 = mayo |> 
 within.data.frame(expr = {
  time = as.difftime(time, units = 'days')
  edp = Surv(time, censor)
 }) |>
 survival_roc(formula = edp ~ mayoscore4, predict.time = 365)
Sprintf.survival_roc(m1)
autoplot.roc(m1)
\dontrun{autoplot(m1)} # unicode error on devtools::check
data(VA, package = 'MASS')
va1 = within(VA, {
 cell = factor(cell)
 tx = (VA$treat == 1)
 status[stime>500] <- 0
 stime[stime>500 ] <- 500
})
library(survival)
fit0 = coxph(Surv(stime,status) ~ Karn + cell + tx + age, data = va1, na.action = na.omit)
va1$eta = fit0$linear.predictor
new = riskset_roc(Surv(stime,status) ~ eta, data = va1, predict.time = 30)
autoplot(new)

}
\references{
\url{https://rpubs.com/kaz_yos/survival-auc}
}
