% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pROC_ext_vanilla.R
\name{get_roc}
\alias{get_roc}
\title{Helper Files of \link[pROC]{roc} Plot}
\usage{
get_roc(formula, data, pctR = 0.6, ...)
}
\arguments{
\item{formula}{\link[stats]{formula} of the univariable or multivariable logistic regression model}

\item{data}{\link[base]{data.frame}}

\item{pctR}{\link[base]{numeric} scalar between 0 and 1,
percentage of training data from input \code{data}
(stratified sampling from response and non-response).
Or use \code{FALSE} to indicate all subjects are used in both training data and test data.
Default \code{.6}.}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\examples{
# ?datasets::infert
# using the wrong model just to illustrate the creation of ROC plot

# will change with each run, because of sampling
set.seed(12); list(
  'with-seed' = get_roc(case ~ spontaneous + induced, data = infert, pctR = .6)
) |> fastmd::render_(file = 'roc1')

# will not change with each run
list(
  'without-seed' = get_roc(case ~ spontaneous + induced, data = infert, pctR = FALSE)
) |> fastmd::render_(file = 'roc2')

}
\references{
Inspired by \url{https://daviddalpiaz.github.io/r4sl/logistic-regression.html}
}
\keyword{internal}
